Added data validation and exception handler service middleware to handle BadHttpRequestExceptions, but this is just working to not stop the programm
Changed the domain model and DTO'sPrice type to decimal to more easy validate and
check it inside the DB. Additionally, migrated the DB for changing the
type of the property of the domain model

+++ Models/DTO/ApartmentDto.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Name { get; set; }
         public uint Rooms { get; set; }
-        public uint Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; }
     }
 }
