+++ Models/DTO/CreateApartmentRequestDto.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public uint Rooms { get; set; }
-        public uint Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; }
     }
 }
